Feat: Defer script & implement round
Deferred the script to make it run after the HTML has loaded

Created a function where two arguments are required,
and based on those two, a winner for RPS is declared!
Then, we alert the message saying the results of the current round
showing who won, and then the score.