Feat: Prepare round function to accept multiple rounds & avoid errors
Now, we display the current round and, if we are already at the 5th one,
we display who won.

We now also keep track of the round, and increment it at the end.

If the user decides to cancel the Choice prompt, then instead of
throwing an error, we handle it properly by saying "That's not a valid
answer" and then sending again the prompt.